Add adt-virt-lxc test runner
Many thanks to Robie Basak for this! This runner provides "revert" and
"revert-full-system" capabilities which is useful for packages which have the
"breaks-testbed" restriction and multiple tests. For the time being this only
works with the ubuntu-cloud template; support for general debian/ubuntu
containers requires an lxc-wait mechanism ("did the container finish booting?")
and getting along with the different user setups (debian using "root",  ubuntu
using an "ubuntu" user with sudo capabilities).