Add support for virtual "@builddeps@" test dependency
This will be replaced with the package's B-D and B-D-I. Document in
doc/README.package-tests. This is useful if you have many build
dependencies which are only necessary for running the test suite and you
don't want to replicate them in the test Depends:. However, please use
this sparingly, as this can easily lead to missing binary package
dependencies being overlooked if they get pulled in via build
dependencies.

(Closes: #720458)