Add first working version of adt-virt-qemu
This uses nbd to put an rc2.d init script into the overlay that opens a root
shell. This gets us a foot into the door to control operations with the guest.

This uses a virtfs shared folder to exchange data with the VM, to avoid having
to squeeze lots of data through a pty:

  http://wiki.qemu.org/Documentation/9psetup

TODO:
 - The nbd approach requires root
 - Make more efficient use of the shared folder in copyup/copydown
 - Add support for putting images and options into a config file