Add downtmp-host testbed capability
Introduce a new "downtmp-host=<path>" testbed capability. If the testbed has
the ability of setting up a shared directory with the host, this gives the host
directory path of the "downtmp" directory as reported by the virt server's
hook_downtmp() function.

This is then used by copyupdown to more efficiently copy files back and forth.
Instead of tar/cat they are copied via the shared tmp dir; in most cases
files are already in there, so they don't need to be (explicitly) copied at all
any more.