Generalize live test output for all runners with shared downtmp
Introduce a new testbed capability "downtmp-shared-fifo" that declares that the
testbed and the host can use a FIFO in the shared downtmp for communication.

If the testbed supports a shared downtmp with FIFOs, use that for getting live
test output. This makes live test output work for all runners except QEMU (9p
does not support FIFOs).