Don't define $TMPDIR for tests any more
Having $TMPDIR in the shared dir isn't efficient and even breaks some tests (as
they might not be able to put sockets etc. into it), and it's an artificial
difference to the default system behaviour of just using /tmp. The
documentation already was changed a long time ago to only document $ADTTMP as a
scratch directory for tests.