Add test restrictions "isolation-container" and "isolation-machine"
Add corresponding testbed capabilities. Tests can use that to declare that they
want to start services or open ports (i. e. a simple chroot/schroot is
insufficient) or access hardware, reboot,  and interact the kernel (where even
a container is insufficient), and will be skipped instead of failing when they
run in a virtualization server which does not provide enough isolation.