Add new options -s/--shell-fail and --shell
For running a shell in the testbed after every (failed) test.  This should make
debugging tests or regressions quite a bit simpler.

This actually spawns a shell for the null, chroot, schroot, and lxc runners,
and shows how to login to the VM for qemu (this will hopefully be improved in
the future, but at least makes debugging possible for now).