Don't run apt-get and setup commands with sending output to /dev/stdout.
When redirecting adt-run's output directly into a file this causes a seek to
start and thus partially overwriting the output file.

Instead, introduce a new Testbed.execute_raw() which relies on auxverb's
promise of DTRT with stdout/stderr, and use that directly.

This also puts the apt-get output back to stdout as it used to be before 2.15.

(Closes: #747041)