Add setup-commands/ubuntu-touch-session
--setup-commands script for installing click, the Ubuntu SDK, some extra
packages for autopilot, and launching Xvfb and an user upstart session.
This approximates the environment of an Ubuntu touch device. You can also run
setup-commands/ro-apt afterwards.

Use this in schroot, lxc, or qemu runner. Everything that's required will be
installed, but as that's a lot of packages it is advisable to run this on a
testbed which has at least ubuntu-sdk-libs already pre-installed.