Add adt-virt-ssh runner
Generic runner for an externally set up testbed which assumes nothing else than
a working ssh connection. This can call a "setup script" to create/configure a
testbed (such as spinning up a cloud VM with nova or setting up SSH on a phone
through ADB). See the manpage for details.

Thanks to Jean-Baptiste Lallement!