Provide override for autopilot_module
Allow the adt-run caller to change the "autopilot_module" from a click
package's manifest by setting $ADT_AUTOPILOT_MODULE. With that one can run an
individual test easily.

This is just a hack for now, so don't document it. We need to find a better way
how to deal with this.