Don't synthesize the no-op 'heavy' perl tests for now
These no-op tests are bundled in the pkg-perl default test control
file so that future expansion will not require source uploads for all
packages. This is not a concern with the synthesized test control file,
which is easy to update centrally when necessary.

So let's save the CPU cycles for something more useful until
the future expansion actually happens.