Fix eternal hang on timeout for lxc --sudo
When executing a command in the testbed times out, don't wait for it after
killing if the auxverb includes "sudo". We can't kill those processes, so we'd
wait forever; we can only hope that they eventually terminate themselves after
we stop the testbed. Fixes non-working timeouts for LXC runner with the --sudo
option.