adt-build-{lxc,lxd}: Run testbed setup scripts also if they are not executable
The scripts in setup-commands/* are installed as data and not as proper
binaries, what makes adt-build-{lxc,lxd} not run any setup, producing
mostly broken containers.

Since the setup script is always piped into a shell instead of executed
directly, it is probably not useful to install it with the execution
bits on anyway.

Closes: #809917