adt-virt-qemu: Only add baseimage to the current boot
Drop updating the initrd in autopkgtest-reboot, as this is very expensive, runs
in the context of the test (thus any output will confuse or even break the
test), and assumes that rebuliding the initrd even works (cf. read-only root).

Instead, add a "prepare_reboot" hook and tell QEMU to remove the added drive
right before rebooting again.

Also put the udev rule into /run instead of /etc, so that it does not leak into
a reboot and does not assume that /etc is writable.