FIX: reprotest: get_all_servers: string formatting
The strings on get_all_servers() would use a negative index to truncate
part of the last filename. As of the previous commit, the relevant name
of the server-name is the name after the prefix (that was correctly
indicated).

Use positive indexing to properly extract the name of the virtual
server. Additionally, add a comma as a parameter to the join to neatly
space all the choices.